Revamped function authorization for security and clarity. - Checking functions centralized in the component helper. - Calling functions in every model with data manipulation potential. - Grouped calls into admin, fm, hr, scheduling and documentation. Moved many parameters to the inheritance hierarchy from function calls. Removed the setting of model 'actions'. Reformatted functions for more logical process flows. Consolidated form validation before processing of data before edit/merge views. Merging resources now requires admin access. Deleting mergeable resources is not not explicitly possible. Replaced generalized variable names with resource specific ones in extending views. Fixes error in the assets table where schedules did not have their parent_id value set.

+++ com_thm_organizer/admin/models/campus.php
@@ -24,6 +24,10 @@ class THM_OrganizerModelCampus extends JModelLegacy
      */
     public function save()
     {
+        if (!THM_OrganizerHelperComponent::allowFMAccess()) {
+            throw new Exception(JText::_('COM_THM_ORGANIZER_403'), 403);
+        }
+
         $data      = JFactory::getApplication()->input->get('jform', [], 'array');
         $dataTable = JTable::getInstance('campuses', 'thm_organizerTable');
 
@@ -46,7 +50,9 @@ class THM_OrganizerModelCampus extends JModelLegacy
      */
     public function delete()
     {
-        require_once JPATH_ROOT . '/media/com_thm_organizer/helpers/component.php';
+        if (!THM_OrganizerHelperComponent::allowFMAccess()) {
+            throw new Exception(JText::_('COM_THM_ORGANIZER_403'), 403);
+        }
 
         return THM_OrganizerHelperComponent::delete('campuses');
     }
