+++ com_thm_organizer/admin/models/color.php
@@ -24,6 +24,10 @@ class THM_OrganizerModelColor extends JModelLegacy
      */
     public function save()
     {
+        if (!THM_OrganizerHelperComponent::isAdmin()) {
+            throw new Exception(JText::_('COM_THM_ORGANIZER_403'), 403);
+        }
+
         $data  = JFactory::getApplication()->input->get('jform', [], 'array');
         $table = JTable::getInstance('colors', 'thm_organizerTable');
 
@@ -37,7 +41,9 @@ class THM_OrganizerModelColor extends JModelLegacy
      */
     public function delete()
     {
-        require_once JPATH_ROOT . '/media/com_thm_organizer/helpers/component.php';
+        if (!THM_OrganizerHelperComponent::isAdmin()) {
+            throw new Exception(JText::_('COM_THM_ORGANIZER_403'), 403);
+        }
 
         return THM_OrganizerHelperComponent::delete('colors');
     }
